movings files to new location and updatign routes
