+++ app/pages/addItems/containers/AddAnotherContainerPage.scala
@@ -14,18 +14,14 @@
  * limitations under the License.
  */
 
-package forms
+package pages.addItems.containers
 
-import javax.inject.Inject
+import pages.QuestionPage
+import play.api.libs.json.JsPath
 
-import forms.mappings.Mappings
-import play.api.data.Form
+case object AddAnotherContainerPage extends QuestionPage[Boolean] {
 
-class ContainerNumberFormProvider @Inject() extends Mappings {
+  override def path: JsPath = JsPath \ toString
 
-  def apply(): Form[String] =
-    Form(
-      "value" -> text("containerNumber.error.required")
-        .verifying(maxLength(100, "containerNumber.error.length"))
-    )
+  override def toString: String = "addAnotherContainer"
 }
