+++ conf/app.routes
@@ -504,12 +504,12 @@ POST       /:lrn/add-items/:itemIndex/previous-references/:referenceIndex/extra-
 GET        /:lrn/add-items/:itemIndex/previous-references/:referenceIndex/change-extra-information                 controllers.addItems.previousReferences.ExtraInformationController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = CheckMode)
 POST       /:lrn/add-items/:itemIndex/previous-references/:referenceIndex/change-extra-information                 controllers.addItems.previousReferences.ExtraInformationController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = CheckMode)
 
-GET        /:lrn/containerNumber                        controllers.ContainerNumberController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-POST       /:lrn/containerNumber                        controllers.ContainerNumberController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-GET        /:lrn/changeContainerNumber                  controllers.ContainerNumberController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
-POST       /:lrn/changeContainerNumber                  controllers.ContainerNumberController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+GET        /:lrn/add-items/:itemIndex/containers/container-number                                                        controllers.addItems.containers.ContainerNumberController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, mode: Mode = NormalMode)
+POST       /:lrn/add-items/:itemIndex/containers/container-number                                                        controllers.addItems.containers.ContainerNumberController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, mode: Mode = NormalMode)
+GET        /:lrn/add-items/:itemIndex/containers/change-container-number                                                 controllers.addItems.containers.ContainerNumberController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, mode: Mode = CheckMode)
+POST       /:lrn/add-items/:itemIndex/containers/change-container-number                                                 controllers.addItems.containers.ContainerNumberController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, mode: Mode = CheckMode)
 
-GET        /:lrn/addAnotherContainer                        controllers.AddAnotherContainerController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-POST       /:lrn/addAnotherContainer                        controllers.AddAnotherContainerController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-GET        /:lrn/changeAddAnotherContainer                  controllers.AddAnotherContainerController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
-POST       /:lrn/changeAddAnotherContainer                  controllers.AddAnotherContainerController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+GET        /:lrn/add-items/:itemIndex/containers/add-another-container                                                   controllers.addItems.containers.AddAnotherContainerController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, mode: Mode = NormalMode)
+POST       /:lrn/add-items/:itemIndex/containers/add-another-container                                                   controllers.addItems.containers.AddAnotherContainerController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, mode: Mode = NormalMode)
+GET        /:lrn/add-items/:itemIndex/containers/change-add-another-container                                            controllers.addItems.containers.AddAnotherContainerController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, mode: Mode = CheckMode)
+POST       /:lrn/add-items/:itemIndex/containers/change-add-another-container                                            controllers.addItems.containers.AddAnotherContainerController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, mode: Mode = CheckMode)
