+++ conf/views/addItems/containers/addAnotherContainer.njk
@@ -27,7 +27,7 @@
           }) }}
         {% endif %}
 
-        <form action="{{ routes.controllers.AddAnotherContainerController.onSubmit(lrn, mode).url }}" method="post">
+        <form action="{{ routes.controllers.addItems.containers.AddAnotherContainerController.onSubmit(lrn, mode).url }}" method="post">
 
           {{ csrf() | safe }}
