+++ test/generators/PageGenerators.scala
@@ -20,6 +20,7 @@ import models.Index
 import org.scalacheck.Arbitrary
 import pages._
 import pages.addItems._
+import pages.addItems.containers.{AddAnotherContainerPage, ContainerNumberPage}
 import pages.addItems.traderDetails._
 import pages.guaranteeDetails.{GuaranteeReferencePage, GuaranteeTypePage}
 import pages.movementDetails.PreLodgeDeclarationPage
