+++ test/generators/UserAnswersEntryGenerators.scala
@@ -23,6 +23,7 @@ import pages._
 import pages.addItems.{CommodityCodePage, ConfirmRemoveItemPage}
 import pages.addItems.traderDetails._
 import pages.addItems._
+import pages.addItems.containers.{AddAnotherContainerPage, ContainerNumberPage}
 import pages.guaranteeDetails.{GuaranteeReferencePage, GuaranteeTypePage}
 import pages.movementDetails.PreLodgeDeclarationPage
 import play.api.libs.json.{JsValue, Json}
